added float support

+++ http_/modules/get_pushes.py
@@ -11,8 +11,8 @@ class Module(BaseModule):
 
     def required_param(self):
         return (
-            RequiredParam('beginning_day', int, 7),
-            RequiredParam('ending_day', int, 0)
+            RequiredParam('beginning_day', float, 7),
+            RequiredParam('ending_day', float, 0)
         )
 
     def get_data(self):
@@ -35,8 +35,8 @@ WHERE
 `pushes`.`date_time`
 BETWEEN :time_begin AND :time_end ;''',
             {
-                'time_begin': current - beginning_day * 86400,
-                'time_end': current - ending_day * 86400
+                'time_begin': int(current - beginning_day * 86400),
+                'time_end': int(current - ending_day * 86400)
             })
 
         kw_result = list()
